pyproject.toml
[tool.poetry]
name = "kw-drf-healthy-habit-tracker"
version = "0.1.0"
description = ""
authors = ["KonstantinMikheev <mikhey17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
pillow = "^11.0.0"
ipython = "^8.29.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
djangorestframework-simplejwt = "^5.3.1"
django-stubs = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
